+++ src/main/java/sleppynavigators/studyupbackend/presentation/chat/dto/response/ChatMessageResponse.java
@@ -1,18 +1,21 @@
-package sleppynavigators.studyupbackend.presentation.chat.dto;
+package sleppynavigators.studyupbackend.presentation.chat.dto.response;
 
 import java.time.LocalDateTime;
+
+import jakarta.validation.constraints.NotBlank;
 import lombok.Builder;
+import org.jetbrains.annotations.NotNull;
 import sleppynavigators.studyupbackend.domain.chat.ChatMessage;
 import sleppynavigators.studyupbackend.domain.chat.SenderType;
 
 @Builder
 public record ChatMessageResponse(
-        String id,
-        Long groupId,
-        Long senderId,
-        SenderType senderType,
-        String content,
-        LocalDateTime timestamp
+        @NotNull String id,
+        @NotNull Long groupId,
+        @NotNull Long senderId,
+        @NotNull SenderType senderType,
+        @NotBlank String content,
+        @NotNull LocalDateTime timestamp
 ) {
     public static ChatMessageResponse from(ChatMessage chatMessage) {
         return ChatMessageResponse.builder()
